+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - (Feature) ArangoBackup create retries and MaxIterations limit
 - (Feature) Add Reason in OOM Metric
 - (Feature) PersistentVolume Inspector
+- (Bugfix) Discover Arango image during ID phase
 
 ## [1.2.27](https://github.com/arangodb/kube-arangodb/tree/1.2.27) (2023-04-27)
 - (Feature) Add InSync Cache